tax-filer-backend/requirements.txt
fastapi
uvicorn[standard]
pydantic>=2.0.0
pydantic-settings
python-dotenv
openai>1
httpx
python-jose
passlib[bcrypt]
